requirements.txt
langchain==0.3.7
langchain-core>=0.3.17,<0.4.0
langchain-community==0.3.7
langchain-openai==0.2.6
openai>=1.52.0
langgraph==0.2.39

streamlit==1.38.0
python-dotenv==1.0.1
pypdf==5.0.0
graphviz==0.20.3
pdfplumber==0.11.0

dspy-ai>=2.4.0
litellm>=1.45.0


